+++ src/main/java/com/github/f4b6a3/uuid/timestamp/DefaultTimestampStrategy.java
@@ -109,7 +109,7 @@ public class DefaultTimestampStrategy extends AbstractSequence implements Timest
 	}
 
 	@Override
-	public long next() {
+	public int next() {
 		if (this.value > maxValue) {
 			this.value = minValue;
 			// (3b) Too many requests
